fabrics: move hostnqn/hostid code to fabrics files
All the hostnqn/hostid related code is only necessary for the fabrics
API, thus move it to the fabrics API and use the correct prefix.
