+++ libnvme/src/libnvme.h.in
@@ -16,7 +16,6 @@ extern "C" {
 #include <nvme/ioctl.h>
 #include <nvme/lib-types.h>
 #include <nvme/lib.h>
-#include <nvme/linux.h>
 #include <nvme/mem.h>
 #include <nvme/nvme-cmds.h>
 #include <nvme/nvme-types.h>
