+++ libnvme/src/nvme/fabrics.h
@@ -31,6 +31,58 @@
 struct libnvmf_context;
 
 /**
+ * libnvmf_generate_hostnqn() - Generate a machine specific host nqn
+ * Return: An nvm namespace qualified name string based on the machine
+ * identifier, or NULL if not successful.
+ */
+char *libnvmf_generate_hostnqn(void);
+
+/**
+ * libnvmf_generate_hostnqn_from_hostid() - Generate a host nqn from
+ * host identifier
+ * @hostid:		Host identifier
+ *
+ * If @hostid is NULL, the function generates it based on the machine
+ * identifier.
+ *
+ * Return: On success, an NVMe Qualified Name for host identification. This
+ * name is based on the given host identifier. On failure, NULL.
+ */
+char *libnvmf_generate_hostnqn_from_hostid(char *hostid);
+
+/**
+ * libnvmf_generate_hostid() - Generate a machine specific host identifier
+ *
+ * Return: On success, an identifier string based on the machine identifier to
+ * be used as NVMe Host Identifier, or NULL on failure.
+ */
+char *libnvmf_generate_hostid(void);
+
+/**
+ * libnvmf_read_hostnqn() - Reads the host nvm qualified name from the config
+ *			      default location
+ *
+ * Retrieve the qualified name from the config file located in $SYSCONFDIR/nvme.
+ * $SYSCONFDIR is usually /etc.
+ *
+ * Return: The host nqn, or NULL if unsuccessful. If found, the caller
+ * is responsible to free the string.
+ */
+char *libnvmf_read_hostnqn(void);
+
+/**
+ * libnvmf_read_hostid() - Reads the host identifier from the config default
+ *			     location
+ *
+ * Retrieve the host idenditifer from the config file located in
+ * $SYSCONFDIR/nvme/. $SYSCONFDIR is usually /etc.
+ *
+ * Return: The host identifier, or NULL if unsuccessful. If found, the caller
+ *	   is responsible to free the string.
+ */
+char *libnvmf_read_hostid(void);
+
+/**
  * libnvmf_trtype_str() - Decode TRTYPE field
  * @trtype: value to be decoded
  *
